Reorg and cleanup.
