+++ src/analysis/retail/shaman/elemental/constants.tsx
@@ -16,4 +16,8 @@ export const CRIT_MULTIPLIER = 2.5;
  */
 export const ON_CAST_BUFF_REMOVAL_GRACE_MS = 50;
 
+/**
+ * Width in percentage that the explanation (the left part with text) takes up
+ * in the guide section.  For use with `ExplanationAndDataSubSection`.
+ */
 export const GUIDE_EXPLANATION_PERCENT_WIDTH = 40;
