+++ src/analysis/retail/shaman/elemental/guide/FlameShockSubSection.tsx
@@ -3,7 +3,12 @@ import CombatLogParser from '../CombatLogParser';
 import { GUIDE_EXPLANATION_PERCENT_WIDTH } from '../constants';
 import { ExplanationAndDataSubSection } from 'interface/guide/components/ExplanationRow';
 
-export const FlameShockGuide = ({ modules }: GuideProps<typeof CombatLogParser>) => {
+/**
+ * The guide sub-section about Flame Shock for Elemental Shamans.
+ * The flameshock module is shared with all shaman specs so this element is the
+ * guide tailored to elemental shamans.
+ */
+export const FlameShockSubSection = ({ modules }: GuideProps<typeof CombatLogParser>) => {
   const explanation = <>TODO</>;
   const data = modules.flameShock.statistic();
