--- src/analysis/retail/shaman/elemental/guide/MaelstromSection.tsx
@@ -1,6 +1,0 @@
-import { GuideProps, SubSection } from 'interface/guide';
-import CombatLogParser from '../CombatLogParser';
-
-export const MaelstromSection = ({ info }: GuideProps<typeof CombatLogParser>) => (
-  <SubSection title="Maelstrom">TODO</SubSection>
-);
